Cargando el estado del mundo

+++ server/includes/character.h
@@ -10,8 +10,8 @@
 
 class Character {
 private:
-    uint8_t head;
-    uint8_t body;
+    uint16_t head;
+    uint16_t body;
     Race race;
     Clase clase;
     Statistics stats;
@@ -21,7 +21,7 @@ public:
     Character& operator=(const Character&) = default;
     Character(Character&&) = default;
     Character& operator=(Character&&) = default;
-    Character(const Race& race, const Clase& clase, uint8_t head, uint8_t body);
+    Character(const Race& race, const Clase& clase, uint16_t head, uint16_t body);
 
     const Statistics& getStatistics() const;
 
@@ -29,8 +29,8 @@ public:
     TypeRace getTypeRace() const;
     TypeClase getTypeClase() const;
 
-    const uint8_t& getTypeHead() const;
-    const uint8_t& getTypeBody() const;
+    const uint16_t& getTypeHead() const;
+    const uint16_t& getTypeBody() const;
 
     const uint16_t& getHpFactorRace();
     const uint16_t& getHpFactorClase();
